Check if crossbuild is being sourced and do nothing.
This enables applications to source crossbuild and
use it's functions without triggering any of it's
usages.

When crossbuild is sourced any arguments are ignored.
For developers only.

Moved dependency calculation to generatedepmake()

publishdepmake takes recipesum as an argument, not global variable

Replaced tabs and spaces mix with spaces in the new functions

Replaced hardcoded relative states path

Package collision warning moved to a function

Refactored recipesum to it's own function